Relates to gh-173. Popup fixes and tests.
- Renamed popup-spec to popup-in-window-spec as it is testing popups in a new window
- Renamed popup-spec1 popup-spec
- Minor change to the assert condition in popup-in-window-test.
