+++ test/ui/popup-in-window-test/popup-in-window-test.html
@@ -12,8 +12,8 @@
     <script type="text/montage-serialization">
 {
     "test": {
-        "module": "ui/popuptest/popuptest",
-        "name": "popupTest",
+        "module": "ui/popup-in-window-test/popup-in-window-test",
+        "name": "PopupInWindowTest",
         "properties": {
             "testPopup": {"@": "testPopup"},
             "popupButton": {"@": "popupButton"}
@@ -63,8 +63,8 @@
 <body>
     <!-- TODOz: need to come up with some content here -->
     <div id="popupButton" class="popupButton">PopupButton</div>
-    <section id="testPopup" class="testPopup" style="display: none;">
-        <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec vitae ligula ligula, eget tempus enim. Vivamus fringilla sagittis vehicula. Ut viverra libero ut dolor dapibus ac dignissim velit aliquam. Nullam vitae odio ac tortor sagittis adipiscing eu eu ipsum. Vivamus interdum consectetur tincidunt. Sed fringilla, massa sit amet porttitor sodales, felis velit laoreet ante, ac dictum enim lacus eu augue. Fusce eros ante, cursus nec auctor et, viverra posuere orci. Morbi adipiscing aliquet felis eu convallis. Curabitur vel erat mi. Morbi non est nisl. Curabitur nec massa in risus faucibus laoreet vel at tortor. Vivamus at lectus dui, vel posuere odio. Phasellus condimentum tristique pellentesque. Suspendisse sit amet malesuada nibh.</p>
+    <section id="testPopup" class="testPopup" style="display: none;height: 100px; width: 200px; overflow: none;">
+        <p>Lorem ipsum dolor sit amet.</p>
     </section>
 </body>
 </html>
