+++ test/ui/popup-test/popup-test.html
@@ -12,8 +12,8 @@
     <script type="text/montage-serialization">
 {
     "test": {
-        "module": "ui/popuptest/popuptest",
-        "name": "popupTest",
+        "module": "ui/popup-test/popup-test",
+        "name": "PopupTest",
         "properties": {            
             "popupContent": {"@": "popupContent"},
             "popupButton": {"@": "popupButton"},
